dns: record type enumeration
Currently it's redundant (because the Record* types are designed to be
type-switched over), but requiring that in the Record interface
future-proofs it: if the number of supported record types explodes, a
DNS server implementation might want to use table lookup instead of type
switch, and if some of the Record implementations are in external
packages, it's not possible to retrofit the Type method.  Even if it's
never needed, it's a minor burden; the DNS record types are standardised.